Question 1
Program that gets as input a color image, performs linear scaling in the Luv domain, and writes the
scaled image as output.
Pixel values outside the window should not be changed. Only pixels within the window should be changed.
The scaling in Luv should stretch only the luminance values. You are asked to apply linear scaling that
would map the smallest L value in the specified window to 0, and the largest L value in the specified
window to 100.