Question 3
This is the same as the first program, except that the scaling is to be performed in the xyY domain. The
scaling should stretch only the luminance (Y) values. In the specified window perform linear scaling that
would map the smallest Y value to 0 and the largest Y value to 1.